+++ tests/test_testing.py
@@ -5,6 +5,7 @@ import threading
 import time
 
 import pytest
+from smb.base import SMBTimeout
 
 import pyndows
 from pyndows.testing import samba_mock, SMBConnectionMock
@@ -17,6 +18,16 @@ def test_remaining_files_to_retrieve_when_reset():
     assert (
         str(exception_info.value)
         == "Expected files were not retrieved: {'tests': 'Test'}"
+    )
+
+
+def test_remaining_storeFile_exceptions_when_reset():
+    pyndows.testing.SMBConnectionMock.storeFile_exceptions.append(SMBTimeout)
+    with pytest.raises(Exception) as exception_info:
+        pyndows.testing.SMBConnectionMock.reset()
+    assert (
+        str(exception_info.value)
+        == "storeFile exceptions were not triggered: [<class 'smb.base.SMBTimeout'>]"
     )
 
 
@@ -42,7 +53,7 @@ def test_non_text_file_can_be_stored(samba_mock: SMBConnectionMock, tmpdir):
         distant_file.write(b"Test Content Move")
 
     pyndows.move(
-        connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file"),
+        connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file")
     )
 
     assert (
@@ -61,7 +72,7 @@ def test_async_retrieval(samba_mock: SMBConnectionMock, tmpdir):
     def add_with_delay(delay: int):
         time.sleep(delay)
         pyndows.move(
-            connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file"),
+            connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file")
         )
 
     threading.Thread(target=add_with_delay, args=(2,)).start()
@@ -84,7 +95,7 @@ def test_async_retrieval_timeout(samba_mock: SMBConnectionMock, tmpdir):
     def add_with_delay(delay: int):
         time.sleep(delay)
         pyndows.move(
-            connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file"),
+            connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file")
         )
 
     threading.Thread(target=add_with_delay, args=(2,)).start()
@@ -162,7 +173,7 @@ def test_retrieval_of_stored_non_text_file(samba_mock: SMBConnectionMock, tmpdir
         distant_file.write(b"Test Content Move")
 
     pyndows.move(
-        connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file"),
+        connection, "TestShare", "TestFilePath", os.path.join(tmpdir, "local_file")
     )
 
     pyndows.get(